+++ src/collective/casestudy/interfaces.py
@@ -14,7 +14,11 @@ class ICaseStudySettings(Interface):
     industries = schema.List(
         title=_("Industries."),
         description=_(
-            "help_industries", default="List of industries available for Case Studies"
+            "help_industries",
+            default=(
+                "List of industries available for Case Studies"
+                "Use the format key|value"
+            ),
         ),
         value_type=schema.TextLine(),
         default=[],
@@ -22,7 +26,10 @@ class ICaseStudySettings(Interface):
 
     usages = schema.List(
         title=_("Usage."),
-        description=_("help_usages", default="Categories of Plone usage"),
+        description=_(
+            "help_usages",
+            default=("Categories of Plone usages. " "Use the format key|value"),
+        ),
         value_type=schema.TextLine(),
         default=[],
     )
@@ -30,6 +37,19 @@ class ICaseStudySettings(Interface):
     versions = schema.List(
         title=_("Plone Versions."),
         description=_("help_versions", default="List of Plone Versions"),
+        value_type=schema.TextLine(),
+        default=[],
+    )
+
+    services = schema.List(
+        title=_("Provided Services."),
+        description=_(
+            "help_services",
+            default=(
+                "List of services a provider offers to their customers. "
+                "Use the format key|value"
+            ),
+        ),
         value_type=schema.TextLine(),
         default=[],
     )