+++ src/collective/casestudy/testing.py
@@ -1,7 +1,9 @@
 from plone.app.testing import applyProfile
+from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
 from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import PloneSandboxLayer
+from plone.restapi.testing import PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
 
 import collective.casestudy
 
@@ -17,10 +19,16 @@ class CaseStudyLayer(PloneSandboxLayer):
         applyProfile(portal, "collective.casestudy:default")
 
 
-CASESTUDY_FIXTURE = CaseStudyLayer()
+FIXTURE = CaseStudyLayer()
 
 
-CASESTUDY_INTEGRATION_TESTING = IntegrationTesting(
-    bases=(CASESTUDY_FIXTURE,),
+INTEGRATION_TESTING = IntegrationTesting(
+    bases=(FIXTURE,),
     name="CaseStudyLayer:IntegrationTesting",
+)
+
+
+FUNCTIONAL_TESTING = FunctionalTesting(
+    bases=(FIXTURE, PLONE_RESTAPI_DX_FUNCTIONAL_TESTING),
+    name="CaseStudyLayer:FunctionalTesting",
 )