+++ src/collective/casestudy/vocabularies/industries.py
@@ -11,5 +11,7 @@ def industries_vocabulary(context):
     terms = []
     industries = api.portal.get_registry_record("casestudy.industries")
     for title in industries:
-        terms.append(SimpleTerm(title, title, title))
+        if "|" in title:
+            token, title = title.split("|")
+        terms.append(SimpleTerm(token, token, title))
     return SimpleVocabulary(terms)