+++ tests/vocabularies/test_vocab_organization.py
@@ -4,9 +4,9 @@ from zope.schema.vocabulary import SimpleVocabulary
 import pytest
 
 
-class TestVocabIndustries:
+class TestVocabOrganizationSize:
 
-    name = f"{PACKAGE_NAME}.vocabulary.industries"
+    name = f"{PACKAGE_NAME}.vocabulary.organization_size"
 
     @pytest.fixture(autouse=True)
     def _vocab(self, get_vocabulary, portal):
@@ -19,11 +19,10 @@ class TestVocabIndustries:
     @pytest.mark.parametrize(
         "token",
         [
-            "gov",
-            "edu",
-            "corporate",
-            "media",
-            "ngo",
+            "mei",
+            "small",
+            "medium",
+            "large",
         ],
     )
     def test_token(self, token):