+++ CHANGELOG.md
@@ -3,10 +3,16 @@
 1.0.0a3 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Add Provider content type #2 [ericof]
+
+- Add Providers behavior [ericof]
+
+- Add catalog indexes for providers, country, services [ericof]
+
+- Update pt_BR translation [ericof]
 
 
-1.0.0a2 (2023-05-15)
+## 1.0.0a2 (2023-05-15)
 
 - Added Spanish translation [macagua]